requirements.txt
streamlit==1.17.0
pymongo==4.3.3
requests==2.28.1
plotly==5.7.0
python-dotenv==1.0.0
pandas==1.5.3
